+++ frontend/src/components/PostProductForm/FormImgList/index.js
@@ -1,9 +1,9 @@
 import './style.scss';
-import { createElement } from '../../utils/dom';
-import pictureIcon from '../../../public/assets/postPage/pictureIcon.svg';
-import imgDeleteButton from '../../../public/assets/postPage/imgDeleteButton.svg';
-import { getState, setState, subscribe } from '../../utils/globalObserver';
-import { uploadedImgState } from '../../store/postPage';
+import { createElement } from '../../../utils/dom';
+import pictureIcon from '../../../../public/assets/postPage/pictureIcon.svg';
+import imgDeleteButton from '../../../../public/assets/postPage/imgDeleteButton.svg';
+import { getState, setState, subscribe } from '../../../utils/globalObserver';
+import { uploadedImgState } from '../../../store/postPage';
 
 export default class FormImgList {
   constructor() {
