+++ frontend/src/components/PostProductForm/FormImgList/style.scss
@@ -1,0 +1,51 @@
+@import '../../../style/variables.scss';
+
+.form-img-list {
+  display: flex;
+  overflow-x: scroll;
+  gap: 16px;
+  &::-webkit-scrollbar {
+    display: none;
+  }
+
+  input {
+    display: none;
+  }
+}
+
+.form-img {
+  position: relative;
+  & > img {
+    width: 76px;
+  }
+
+  &-delete-btn {
+    position: absolute;
+    top: -10px;
+    right: -10px;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    width: 24px;
+    height: 24px;
+    background-color: $black;
+    border-radius: 50%;
+  }
+}
+
+.form-img-add-btn {
+  min-width: 76px;
+  height: 76px;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  background-color: $off-white;
+  border: 1px solid $gray3;
+  border-radius: 8px;
+  .form-img-count {
+    font-size: 12px;
+    color: $gray1;
+    margin-top: 7px;
+  }
+}
